refactor: remove unnecessary class
