--- src/main/java/nextstep/subway/util/DiscountPolicy.java
@@ -1,7 +1,0 @@
-package nextstep.subway.util;
-
-@FunctionalInterface
-public interface DiscountPolicy {
-
-    int discount(int fare);
-}
